poprawka która wyłącza próbę kopiowania katalogu w iteracjach pętli for

+++ 06-for_loops-2a.sh
@@ -10,5 +10,9 @@ fi
 
 for file in $(ls $current_dir)
 do
-    cp "$current_dir/$file" "$backup_dir/$prefix$file"
-    # echo "$prefix$file"
+    if [ -f $file ]; then
+        cp "$current_dir/$file" "$backup_dir/$prefix$file"
+        # echo "$prefix$file"
+    fi
+
+done
